Moved logic from the webserver class to the node class. In the same way that the multiwallet is just a front end for the wallet logic, the webserver should be the front end for the node logic
